Actually put reviews.io component in desing system components
