+++ es-vue-base/src/lib-components/index.js
@@ -18,6 +18,7 @@ export { default as EsProgress } from './EsProgress.vue';
 export { default as EsRating } from './EsRating.vue';
 export { default as EsReview } from './EsReview.vue';
 export { default as EsReviewModal } from './EsReviewModal.vue';
+export { default as EsReviewsIoCardCarousel } from './EsReviewsIoCardCarousel.vue';
 export { default as EsSlider } from './EsSlider.vue';
 export { default as EsSupport } from './EsSupport.vue';
 export { default as EsTab } from './EsTab.vue';
